Fix ListVPCs method for AWI Domain Connector
Awi-CLI and Awi-UI use awi-infra-guard codebase for listing most
resources including VPCs. However, kube-awi uses older API from
awi-grpc-catalyst-sdwan rather than awi-infra-guard.

Currently, most of methods from awi-grpc-catalyst-sdwan are
forwarded to awi-infra-guard anyway, but ListVPCs is not and it
was implemented to return empty list of VPCs if the connector was
not VManage.

This change fixes it by using awi-infra-guard if the connector is
different than VManage.